% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multideploy-package.R
\docType{package}
\name{multideploy-package}
\alias{multideploy}
\alias{multideploy-package}
\title{This package provides functions to deploy file changes across
multiple GitHub repositories using the gh package.}
\description{
Deploy file changes across multiple 'GitHub' repositories using the 'GitHub' 'Web API' \url{https://docs.github.com/en/rest}. Allows synchronizing common files, Continuous Integration ('CI') workflows, or configurations across many repositories with a single command.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-pkg.thecoatlessprofessor.com/multideploy/}
  \item \url{https://github.com/coatless-rpkg/multideploy}
  \item Report bugs at \url{https://github.com/coatless-rpkg/multideploy/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{james.balamuta@gmail.com} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID})

}
\keyword{internal}
