% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-mapping.R
\name{print.file_mapping}
\alias{print.file_mapping}
\title{Print method for file_mapping objects}
\usage{
\method{print}{file_mapping}(x, max_files = 20, ...)
}
\arguments{
\item{x}{A file_mapping object to print}

\item{max_files}{Maximum number of files to display. Default is 20.}

\item{...}{Additional arguments passed to print methods (not used)}
}
\value{
Invisibly returns the file_mapping object
}
\description{
Print method for file_mapping objects
}
