% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_content}
\alias{file_content}
\title{Retrieve the content of a file from a GitHub repository}
\usage{
file_content(repo, path, ref = NULL, quiet = FALSE)
}
\arguments{
\item{repo}{Character string specifying the full name of the repository (owner/repo)}

\item{path}{Character string specifying the path to the file}

\item{ref}{Character string specifying the branch name or commit SHA. Default is NULL.}
}
\value{
List with the content and SHA of the file if it exists, NULL otherwise
}
\description{
Retrieve the content of a file from a GitHub repository
}
\examples{
\dontrun{
file_info <- file_content("username/repo", "path/to/file.R")
}
}
