% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multideploy.R
\name{orgs}
\alias{orgs}
\title{List organizations for the authenticated user}
\usage{
orgs(per_page = 100, max_pages = 5, quiet = FALSE)
}
\arguments{
\item{per_page}{Number of organizations to return per page. Default is 100.}

\item{max_pages}{Maximum number of pages to retrieve. Default is 5.}
}
\value{
A data frame of organizations with their login names
}
\description{
List organizations for the authenticated user
}
\examples{
\dontrun{
organizations <- orgs()
}
}
