% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{print.pr_create_result}
\alias{print.pr_create_result}
\title{Print method for pr_create_result objects}
\usage{
\method{print}{pr_create_result}(x, ...)
}
\arguments{
\item{x}{The pr_create_result object to print}

\item{...}{Additional arguments to pass to print methods}
}
\value{
Invisibly returns the input data frame
}
\description{
Print method for pr_create_result objects
}
\examples{
\dontrun{
results <- pr_create(repos, "branch", "title", "body", file_mapping)
print(results)
}
}
